+++ tests/testthat/setup-add-colnames.R
@@ -1,0 +1,1 @@
+oo <- options(huxtable.add_colnames = FALSE)